Use float instead of double for GNN inputs
Changed tensor type from double to float for better GPU compatibility.

+++ code/model.py
@@ -86,7 +86,8 @@ class GNNReGVD(nn.Module):
     def forward(self, adj=None, adj_mask=None, adj_feature=None, labels=None):
         # run over GNNs
         adj = torch.squeeze((adj))
-        outputs = self.gnn(adj_feature.to(device).double(), adj.to(device).double(), adj_mask.to(device).double())
+        # Use float() instead of double() for better GPU compatibility
+        outputs = self.gnn(adj_feature.to(device).float(), adj.to(device).float(), adj_mask.to(device).float())
         logits = self.classifier(outputs)
         prob = F.sigmoid(logits)
         if labels is not None:
@@ -98,3 +99,4 @@ class GNNReGVD(nn.Module):
             return prob
 
 
+
